ch(mail+inquiries): send email automatically, implement inquiries

+++ backend/app/api/v2/product/delete.py
@@ -7,12 +7,19 @@ from app.db import get_db
 
 router = APIRouter()
 
+not_found = HTTPException(
+    status_code=status.HTTP_404_NOT_FOUND,
+    detail="Product not found!"
+)
+
 
 @router.delete("/{product_id}", response_model=Res, dependencies=[Depends(get_current_user)])
 async def delete_product(
         product_id: int = Path(title="Product ID", description="The id of the product to be deleted"),
         db: Session = Depends(get_db)) -> Res:
     product = db.query(Product).filter_by(id=product_id).first()
+    if not product:
+        raise not_found
     db.delete(product)
     db.commit()
